refactor: remove main structs to mod.rs
